Add styles for show talk component
