+++ src/components/style.css
@@ -1,6 +1,4 @@
-#root{
-    background-color: rgb(0, 0, 128);
-}
+
 input{
     color: #000080;
     margin: 5px;
@@ -57,4 +55,48 @@ button{
 .speakerLocation{
     display: flex;
     justify-content: space-evenly;
+}
+
+.showTalk{
+    background-color: white;
+    margin: 0 auto;
+    width: 90%;
+    padding: 10px;
+    border-radius: 10px;
+}
+.showDate{
+    display: flex;
+    justify-content: space-between;
+}
+.talkDesc{
+    background-color: white;
+}
+
+.talkDesc{
+    width: 85%;
+    margin: 10px auto;
+}
+.desc{
+    font-size: 15px;
+    font-weight: bold;
+    color: rgb(0, 0, 128);
+    opacity: 0.7;
+    margin-bottom: 3px;
+}
+
+.talkSpeakers{
+    background-color: #F8F8FF;
+    padding: 20px;
+    margin: 20px;
+}
+.talkSpeakers p{
+    color: green;
+}
+.speaker{
+    color: blue;
+    padding: 10px;
+    background-color: white;
+    border: 1px solid lightgray;
+    border-radius: 5px;
+    margin: 5px;
 }
